requirements.txt
openai>=1.12.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
PyPDF2>=3.0.1
tqdm>=4.66.1
tenacity>=8.2.0
streamlit>=1.31.0
